refactor: Aligns service and tests names to recent changes.
