+++ src/Modules/Courses/Tutor.Courses.Tests/Integration/Supervision/ActiveSupervisionTests.cs
@@ -7,12 +7,12 @@ using Tutor.Courses.API.Dtos.Groups;
 using Tutor.Courses.API.Dtos.Reflections;
 using Tutor.Courses.API.Public.Supervision;
 
-namespace Tutor.Courses.Tests.Integration.Monitoring;
+namespace Tutor.Courses.Tests.Integration.Supervision;
 
 [Collection("Sequential")]
-public class CourseMonitoringTests : BaseCoursesIntegrationTest
+public class ActiveSupervisionTests : BaseCoursesIntegrationTest
 {
-    public CourseMonitoringTests(CoursesTestFactory factory) : base(factory) {}
+    public ActiveSupervisionTests(CoursesTestFactory factory) : base(factory) {}
 
     [Fact]
     public void GetActiveCourses_ReturnsActiveCourses()
@@ -175,7 +175,7 @@ public class CourseMonitoringTests : BaseCoursesIntegrationTest
 
     private static ActiveSupervisionController CreateController(IServiceScope scope)
     {
-        return new ActiveSupervisionController(scope.ServiceProvider.GetRequiredService<ICourseMonitoringService>())
+        return new ActiveSupervisionController(scope.ServiceProvider.GetRequiredService<IActiveSupervisionService>())
         {
             ControllerContext = BuildContext("0", "administrator")
         };
